move Timeline.tsx into course components folder
